Add files via upload
This is the final product of the algorithm. Not particularly impressive but took a lot of work. This video showcases the importance of the smoothing algorithm that was run toward the end of the program. The previous video in the repository shows the same video but without the smoothing algorithm.